ocl: prepared OpenCL backend for CP2K's Offload/DBM/DBT libraries
* Introduced ACC_OPENCL_MEM_DEBUG for some debug details.
* Auto-detecting device-memory offset (ACC_OPENCL_MEM_OFFSET), i.e.,
  range-finding device pointers to adjust the offset accordingly.
  - c_dbcsr_acc_memcpy_d2d, c_dbcsr_acc_memcpy_h2d,
  - c_dbcsr_acc_memcpy_d2h, c_dbcsr_acc_opencl_memset.
* Resolved cast-qual warnings.
* Removed ACC_OPENCL_MEM_NOALLOC/ACC_OPENCL_MEM macros.
* Improved performance when using clEnqueueFillBuffer.
* Allow stream-argument (ACC-interface) to be NULL
  - Introduced c_dbcsr_acc_opencl_stream_default().
  - Implemented ACC_OPENCL_STREAM_NULL.
  - Enables synchronous operation).
* Exposed/implemented OpenCL-only c_dbcsr_acc_opencl_memset,
  and based c_dbcsr_acc_memset_zero on c_dbcsr_acc_opencl_memset.
  - Allow custom memset-value to match CP2K's offload library
* Adjusted console output (ACC_OPENCL_VERBOSE).
* Revised acc_opencl.sh (generating kernel header).
* Increased number of repetitions of standalone benchmark.
* Updated tuned parameters.

+++ src/acc/opencl/acc_opencl.sh
@@ -37,8 +37,7 @@ then
     -h|--help)
       shift $#;;
     -p|--params)
-      PARAMPATH=yes
-      PARAMS=$2
+      PARAMS="$2\t"
       shift 2;;
     -c|-d|--debug|--comments)
       CPPFLAGS+=" -C"
@@ -50,13 +49,9 @@ then
     esac
   done
   HERE="$(cd "$(dirname "$0")" && pwd -P)"
-  PARAMDIR=$(if [ ! "${PARAMDIR}" ]; then echo "${HERE}/smm/params"; fi)
-  if [ "${PARAMPATH}" ]; then
-    PARAMPATH=${PARAMS}
-  else
-    HERE="$(cd "$(dirname "$0")" && pwd -P)"
-    PARAMPATH=${PARAMDIR}
-  fi
+  PARAMDIR=${PARAMDIR:-${PARAMS}}
+  PARAMDIR=${PARAMDIR:-${HERE}/smm/params}
+  PARAMDIR=$(echo -e "${PARAMDIR}" | ${TR} -d '\t')
   if [ "$#" -gt 1 ]; then
     # allow for instance /dev/stdout
     if [ "${OFILE##*.}" = "h" ]; then
@@ -127,7 +122,7 @@ then
     NFILES_CSV=0
     for CSVFILE in "${CSVFILES[@]}"; do
       if [ "${CSVFILE##*.}" = "csv" ]; then
-        if [ -e "${CSVFILE}" ]; then
+        if [ -f "${CSVFILE}" ]; then
           NFILES_CSV=$((NFILES_CSV+1))
         fi
       else
@@ -136,15 +131,15 @@ then
         exit 1
       fi
     done
-    if [ "0" = "${NFILES_CSV}" ] && [ "${PARAMPATH}" ]; then
-      CSVFILES=("${PARAMPATH}"/*.csv)
+    if [ "0" = "${NFILES_CSV}" ] && [ "${PARAMDIR}" ] && [ -d "${PARAMDIR}" ]; then
+      CSVFILES=("${PARAMDIR}"/*.csv)
       NFILES_CSV=${#CSVFILES[@]}
     fi
     for CSVFILE in "${CSVFILES[@]}"; do
       if [ ! "${DELIM}" ]; then
-        SEPAR=$(${SED} -n "1s/[^${DELIMS}]//gp" "${CSVFILE}")
+        SEPAR=$(${SED} -n "1s/[^${DELIMS}]//gp" "${CSVFILE}" 2>/dev/null)
         DELIM=${SEPAR:0:1}
-        MATCH=$(${SED} -n "1s/[^${DELIM}]//gp" "${CSVFILE}")
+        MATCH=$(${SED} -n "1s/[^${DELIM}]//gp" "${CSVFILE}" 2>/dev/null)
       fi
       if [ "${DELIM}" ]; then
         CHECK=$(${SED} "/^[[:space:]]*$/d;s/[^${DELIM}]//g" "${CSVFILE}" | ${SORT} -u | ${SED} -n "0,/./p")
@@ -156,7 +151,7 @@ then
       else
         ERRFILE=${CSVFILE}
       fi
-      if [ "${ERRFILE}" ]; then
+      if [ "${ERRFILE}" ] && [ -f "${ERRFILE}" ]; then
         >&2 echo "ERROR: ${ERRFILE} is malformed!"
         if [ "${HFILE}" ]; then ${RM} -f "${OFILE}"; fi
         exit 1
